refactor(api): Separate mutable configuration
