+++ unchained-api/src/main/java/unchained/Configuration.java
@@ -4,10 +4,8 @@ import java.util.Set;
 
 /**
  * This interface encapsulates a malleable configuration which contains arbitrary values.
- *
- * @param <C> the configuration type.
  */
-public interface Configuration<C extends Configuration<C>> {
+public interface Configuration {
 
     /**
      * This interface represents a configuration option definition, which lets us do things like statically binding
@@ -110,46 +108,6 @@ public interface Configuration<C extends Configuration<C>> {
     }
 
     /**
-     * Enables the provided key by setting it to {@code true}.
-     *
-     * @param key the key to enable.
-     * @return the configuration itself to use for chaining method calls.
-     */
-    default C enable(String key) {
-        return set(key, true);
-    }
-
-    /**
-     * Enables the provided option by setting it to {@code true}.
-     *
-     * @param option the option to enable.
-     * @return the configuration itself to use for chaining method calls.
-     */
-    default C enable(Option<?> option) {
-        return set(option, true);
-    }
-
-    /**
-     * Disable the provided key by setting it to {@code false}.
-     *
-     * @param key the key to disable.
-     * @return the configuration itself to use for chaining method calls.
-     */
-    default C disable(String key) {
-        return set(key, false);
-    }
-
-    /**
-     * Disable the provided option by setting it to {@code false}.
-     *
-     * @param option the option to disable.
-     * @return the configuration itself to use for chaining method calls.
-     */
-    default C disable(Option<?> option) {
-        return set(option, false);
-    }
-
-    /**
      * Returns the value of the given key, or {@code null} if no such key exists.
      *
      * @param key the key.
@@ -187,49 +145,4 @@ public interface Configuration<C extends Configuration<C>> {
      */
     <E> E get(String key, E defaultValue);
 
-    /**
-     * Changes the value of the provided key.
-     *
-     * @param key the key.
-     * @param value the value.
-     * @return the configuration itself to use for chaining method calls.
-     */
-    C set(String key, Object value);
-
-    /**
-     * Sets the value of the provided option to the given value.
-     *
-     * @param option the option.
-     * @param value the new value.
-     * @param <E> the type of the value.
-     * @return the configuration itself to use for chaining method calls.
-     */
-    default <E> C set(Option<E> option, Object value) {
-        return set(option.key(), value);
-    }
-
-    /**
-     * Removes the value of the provided key from the configuration and returns its previous value if it was ever set.
-     *
-     * @param key the key to unset.
-     * @param <E> the type of the option.
-     * @return the previously configured value of the configuration option, or {@code null} if it was never set.
-     */
-    <E> E unset(String key);
-
-    /**
-     * <p>Removes the value of the provided option from the configuration and returns its previous value if it was ever
-     * set.</p>
-     *
-     * <pThis method does not provide any sort of deserialization. Type-casting is done via coercion; i.e. if a value
-     * is stored using a type, but read using another, incompatible type, an exception will occur.</p>
-     *
-     * @param option the option to unset.
-     * @param <E> the type of the option.
-     * @return the previously configured value of the configuration option, or {@code null} if it was never set.
-     */
-    default <E> E unset(Option<?> option) {
-        return unset(option.key());
-    }
-
-}
+ }
